demolearn-39[Feat]: 결제 완료 ui & particles background & 케러셀 수정
