+++ app/page.tsx
@@ -5,11 +5,11 @@ import StudentReviews from "@/components/main/StudentReviews";
 
 export default function Home() {
   return (
-    <main className="min-h-screen bg-white">
+    <>
       <HeroSection />
       <CourseSection title="지금 진행중인 챌린지" data={challenges} />
       <CourseSection title="지금 가장 주목받는 강의" data={courses} />
       <StudentReviews />
-    </main>
+    </>
   );
 }
