+++ app/payment/[id]/page.tsx
@@ -1,10 +1,16 @@
-"use client";
+import { Suspense } from "react";
+import PaymentCompletePage from "@/components/payment/PaymentCompletePage";
 
-export default function PaymentSuccessPage() {
-    <div className="min-h-screen bg-gray-50 py-12">
-      <div className="max-w-4xl mx-auto px-4">
-        <h1 className="text-3xl font-bold text-gray-900 mb-8">결제 페이지</h1>
-        <p className="text-gray-600">결제 페이지가 곧 구현될 예정입니다.</p>
+export default function PaymentSuccess() {
+  return (
+    <Suspense fallback={
+      <div className="min-h-screen bg-gray-50 py-12">
+        <div className="max-w-4xl mx-auto px-4">
+          <p className="text-gray-600">로딩 중...</p>
+        </div>
       </div>
-    </div>
+    }>
+      <PaymentCompletePage />
+    </Suspense>
+  );
 }
