+++ package.json
@@ -14,11 +14,14 @@
     "clsx": "^2.1.1",
     "lucide-react": "^0.525.0",
     "next": "15.3.5",
+    "ogl": "^1.0.11",
     "react": "^19.1.1",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
     "react-multi-carousel": "^2.8.6",
-    "tailwind-merge": "^3.3.1"
+    "react-tsparticles": "^2.12.2",
+    "tailwind-merge": "^3.3.1",
+    "tsparticles-slim": "^2.12.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
